move heuristics to dedicated folder
