+++ src/main/java/com/mozilla/secops/httprequest/HTTPRequest.java
@@ -1,18 +1,19 @@
 package com.mozilla.secops.httprequest;
 
 import com.mozilla.secops.DetectNat;
-import com.mozilla.secops.DocumentingTransform;
-import com.mozilla.secops.FileUtil;
 import com.mozilla.secops.IOOptions;
-import com.mozilla.secops.IprepdIO;
 import com.mozilla.secops.OutputOptions;
 import com.mozilla.secops.SourceCorrelation;
-import com.mozilla.secops.Stats;
 import com.mozilla.secops.alert.Alert;
 import com.mozilla.secops.alert.AlertFormatter;
-import com.mozilla.secops.alert.AlertMeta;
-import com.mozilla.secops.alert.AlertSuppressorCount;
-import com.mozilla.secops.httprequest.HTTPRequestMetrics.HeuristicMetrics;
+import com.mozilla.secops.httprequest.heuristics.EndpointAbuseAnalysis;
+import com.mozilla.secops.httprequest.heuristics.EndpointSequenceAbuse;
+import com.mozilla.secops.httprequest.heuristics.ErrorRateAnalysis;
+import com.mozilla.secops.httprequest.heuristics.HardLimitAnalysis;
+import com.mozilla.secops.httprequest.heuristics.PerEndpointErrorRateAnalysis;
+import com.mozilla.secops.httprequest.heuristics.StatusCodeRateAnalysis;
+import com.mozilla.secops.httprequest.heuristics.ThresholdAnalysis;
+import com.mozilla.secops.httprequest.heuristics.UserAgentBlocklistAnalysis;
 import com.mozilla.secops.input.Input;
 import com.mozilla.secops.input.InputElement;
 import com.mozilla.secops.metrics.CfgTickBuilder;
@@ -24,33 +25,23 @@ import com.mozilla.secops.window.GlobalTriggers;
 import java.io.IOException;
 import java.io.Serializable;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.Default;
 import org.apache.beam.sdk.options.Description;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
-import org.apache.beam.sdk.transforms.Count;
-import org.apache.beam.sdk.transforms.Distinct;
 import org.apache.beam.sdk.transforms.DoFn;
 import org.apache.beam.sdk.transforms.Filter;
 import org.apache.beam.sdk.transforms.Flatten;
-import org.apache.beam.sdk.transforms.GroupByKey;
 import org.apache.beam.sdk.transforms.MapElements;
 import org.apache.beam.sdk.transforms.PTransform;
 import org.apache.beam.sdk.transforms.ParDo;
 import org.apache.beam.sdk.transforms.SerializableFunction;
-import org.apache.beam.sdk.transforms.Values;
 import org.apache.beam.sdk.transforms.windowing.AfterProcessingTime;
 import org.apache.beam.sdk.transforms.windowing.AfterWatermark;
-import org.apache.beam.sdk.transforms.windowing.BoundedWindow;
 import org.apache.beam.sdk.transforms.windowing.FixedWindows;
 import org.apache.beam.sdk.transforms.windowing.Repeatedly;
 import org.apache.beam.sdk.transforms.windowing.Sessions;
@@ -62,11 +53,7 @@ import org.apache.beam.sdk.values.PCollectionTuple;
 import org.apache.beam.sdk.values.PCollectionView;
 import org.apache.beam.sdk.values.TupleTag;
 import org.apache.beam.sdk.values.TupleTagList;
-import org.joda.time.DateTime;
 import org.joda.time.Duration;
-import org.joda.time.Instant;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * {@link HTTPRequest} describes and implements a Beam pipeline for analysis of HTTP requests using
@@ -173,1439 +160,6 @@ public class HTTPRequest implements Serializable {
                                           Duration.standardSeconds(paneFiringDelaySeconds)))))
                   .withAllowedLateness(Duration.ZERO)
                   .accumulatingFiredPanes());
-    }
-  }
-
-  /** Transform for analysis of error rates per client within a given window. */
-  public static class ErrorRateAnalysis extends PTransform<PCollection<Event>, PCollection<Alert>>
-      implements DocumentingTransform {
-    private static final long serialVersionUID = 1L;
-
-    private final Long maxErrorRate;
-    private final String monitoredResource;
-    private final Boolean enableIprepdDatastoreExemptions;
-    private final String iprepdDatastoreExemptionsProject;
-
-    private Logger log;
-
-    /**
-     * Static initializer for {@link ErrorRateAnalysis}
-     *
-     * @param toggles {@link HTTPRequestToggles}
-     * @param enableIprepdDatastoreExemptions True to enable datastore exemptions
-     * @param iprepdDatastoreExemptionsProject Project to look for datastore entities in
-     */
-    public ErrorRateAnalysis(
-        HTTPRequestToggles toggles,
-        Boolean enableIprepdDatastoreExemptions,
-        String iprepdDatastoreExemptionsProject) {
-      maxErrorRate = toggles.getMaxClientErrorRate();
-      monitoredResource = toggles.getMonitoredResource();
-      this.enableIprepdDatastoreExemptions = enableIprepdDatastoreExemptions;
-      this.iprepdDatastoreExemptionsProject = iprepdDatastoreExemptionsProject;
-      log = LoggerFactory.getLogger(ErrorRateAnalysis.class);
-    }
-
-    /** {@inheritDoc} */
-    public String getTransformDoc() {
-      return String.format(
-          "Alert if a single source address generates more than %d 4xx errors in a "
-              + "1 minute window.",
-          maxErrorRate);
-    }
-
-    @Override
-    public PCollection<Alert> expand(PCollection<Event> input) {
-      return input
-          .apply(
-              "isolate client errors",
-              ParDo.of(
-                  new DoFn<Event, String>() {
-                    private static final long serialVersionUID = 1L;
-
-                    @ProcessElement
-                    public void processElement(ProcessContext c) {
-                      Normalized n = c.element().getNormalized();
-                      Integer status = n.getRequestStatus();
-                      if (status == null) {
-                        return;
-                      }
-                      if (n.getSourceAddress() == null) {
-                        return;
-                      }
-                      if (status >= 400 && status < 500) {
-                        c.output(n.getSourceAddress());
-                      }
-                    }
-                  }))
-          .apply(Count.<String>perElement())
-          .apply(
-              "per-client error rate analysis",
-              ParDo.of(
-                  new DoFn<KV<String, Long>, Alert>() {
-                    private static final long serialVersionUID = 1L;
-
-                    @ProcessElement
-                    public void processElement(ProcessContext c, BoundedWindow w) {
-                      if (c.element().getValue() <= maxErrorRate) {
-                        return;
-                      }
-                      Alert a = new Alert();
-                      a.setSummary(
-                          String.format(
-                              "%s httprequest error_rate %s %d",
-                              monitoredResource, c.element().getKey(), c.element().getValue()));
-                      a.setCategory("httprequest");
-                      a.setSubcategory("error_rate");
-                      a.addMetadata(AlertMeta.Key.SOURCEADDRESS, c.element().getKey());
-
-                      if (enableIprepdDatastoreExemptions) {
-                        try {
-                          IprepdIO.addMetadataIfIpIsExempt(
-                              c.element().getKey(), a, iprepdDatastoreExemptionsProject);
-                        } catch (IOException exc) {
-                          log.error("error checking iprepd exemptions: {}", exc.getMessage());
-                          return;
-                        }
-                      }
-
-                      a.addMetadata(AlertMeta.Key.ERROR_COUNT, c.element().getValue().toString());
-                      a.addMetadata(AlertMeta.Key.ERROR_THRESHOLD, maxErrorRate.toString());
-                      a.setNotifyMergeKey(String.format("%s error_count", monitoredResource));
-                      a.addMetadata(
-                          AlertMeta.Key.WINDOW_TIMESTAMP,
-                          (new DateTime(w.maxTimestamp())).toString());
-                      if (!a.hasCorrectFields()) {
-                        throw new IllegalArgumentException("alert has invalid field configuration");
-                      }
-                      c.output(a);
-                    }
-                  }));
-    }
-  }
-
-  /** Transform for analysis of hard per-source request count limit within fixed window */
-  public static class HardLimitAnalysis extends PTransform<PCollection<Event>, PCollection<Alert>>
-      implements DocumentingTransform {
-    private static final long serialVersionUID = 1L;
-
-    private final Long maxCount;
-    private final String monitoredResource;
-    private final Boolean enableIprepdDatastoreExemptions;
-    private final String iprepdDatastoreExemptionsProject;
-    private PCollectionView<Map<String, Boolean>> natView = null;
-    private final HeuristicMetrics metrics;
-
-    private Logger log;
-
-    /**
-     * Static initializer for {@link HardLimitAnalysis}
-     *
-     * @param toggles {@link HTTPRequestToggles}
-     * @param enableIprepdDatastoreExemptions True to enable datastore exemptions
-     * @param iprepdDatastoreExemptionsProject Project to look for datastore entities in
-     * @param natView Use {@link DetectNat} view, or null to disable
-     */
-    public HardLimitAnalysis(
-        HTTPRequestToggles toggles,
-        Boolean enableIprepdDatastoreExemptions,
-        String iprepdDatastoreExemptionsProject,
-        PCollectionView<Map<String, Boolean>> natView) {
-      maxCount = toggles.getHardLimitRequestCount();
-      monitoredResource = toggles.getMonitoredResource();
-      this.enableIprepdDatastoreExemptions = enableIprepdDatastoreExemptions;
-      this.iprepdDatastoreExemptionsProject = iprepdDatastoreExemptionsProject;
-      this.natView = natView;
-      log = LoggerFactory.getLogger(HardLimitAnalysis.class);
-      metrics = new HeuristicMetrics(HardLimitAnalysis.class.getName());
-    }
-
-    /** {@inheritDoc} */
-    public String getTransformDoc() {
-      return String.format(
-          "Alert if single source address makes more than %d requests in a 1 minute window.",
-          maxCount);
-    }
-
-    /**
-     * Function to be used with filter transform to filter out only ip, count pairs where the count
-     * is higher than the hard limit. This is used we only need the side input when an ip address is
-     * over the hard limit.
-     */
-    private class HasCountGreaterThan implements SerializableFunction<KV<String, Long>, Boolean> {
-      private static final long serialVersionUID = 1L;
-
-      @Override
-      public Boolean apply(KV<String, Long> kv) {
-        if (kv.getValue() <= maxCount) {
-          return false;
-        }
-        return true;
-      }
-    }
-
-    @Override
-    @SuppressWarnings("deprecation")
-    public PCollection<Alert> expand(PCollection<Event> input) {
-      if (natView == null) {
-        // If natView was not set then we just create an empty view for use as the side input
-        natView = DetectNat.getEmptyView(input.getPipeline());
-      }
-      return input
-          .apply(
-              "extract client ip",
-              ParDo.of(
-                  new DoFn<Event, String>() {
-                    private static final long serialVersionUID = 1L;
-
-                    @ProcessElement
-                    public void processElement(ProcessContext c) {
-                      Normalized n = c.element().getNormalized();
-                      if (n.getSourceAddress() == null) {
-                        return;
-                      }
-                      c.output(n.getSourceAddress());
-                    }
-                  }))
-          .apply("hard limit per client count", Count.<String>perElement())
-          .apply("filter clients above limit", Filter.by(new HasCountGreaterThan()))
-          // Reshuffle to prevent fusion of steps with large input that do not need side input
-          // with low volume step that needs side input
-          .apply("force materialization of input", org.apache.beam.sdk.transforms.Reshuffle.of())
-          .apply(
-              "per-source hard limit analysis",
-              ParDo.of(
-                      new DoFn<KV<String, Long>, Alert>() {
-                        private static final long serialVersionUID = 1L;
-
-                        @ProcessElement
-                        public void processElement(ProcessContext c, BoundedWindow w) {
-                          Map<String, Boolean> nv = c.sideInput(natView);
-                          Boolean isNat = nv.get(c.element().getKey());
-                          if (isNat != null && isNat) {
-                            log.info(
-                                "{}: detectnat: skipping result emission for {}",
-                                w.toString(),
-                                c.element().getKey());
-                            metrics.natDetected();
-                            return;
-                          }
-                          Alert a = new Alert();
-                          a.setSummary(
-                              String.format(
-                                  "%s httprequest hard_limit %s %d",
-                                  monitoredResource, c.element().getKey(), c.element().getValue()));
-                          a.setCategory("httprequest");
-                          a.setSubcategory("hard_limit");
-                          a.addMetadata(AlertMeta.Key.SOURCEADDRESS, c.element().getKey());
-
-                          try {
-                            if (enableIprepdDatastoreExemptions) {
-                              IprepdIO.addMetadataIfIpIsExempt(
-                                  c.element().getKey(), a, iprepdDatastoreExemptionsProject);
-                            }
-                          } catch (IOException exc) {
-                            log.error("error checking iprepd exemptions: {}", exc.getMessage());
-                            return;
-                          }
-
-                          a.addMetadata(AlertMeta.Key.COUNT, c.element().getValue().toString());
-                          a.addMetadata(AlertMeta.Key.REQUEST_THRESHOLD, maxCount.toString());
-                          a.setNotifyMergeKey(
-                              String.format("%s hard_limit_count", monitoredResource));
-                          a.addMetadata(
-                              AlertMeta.Key.WINDOW_TIMESTAMP,
-                              (new DateTime(w.maxTimestamp())).toString());
-                          if (!a.hasCorrectFields()) {
-                            throw new IllegalArgumentException(
-                                "alert has invalid field configuration");
-                          }
-                          c.output(a);
-                        }
-                      })
-                  .withSideInputs(natView));
-    }
-  }
-
-  /** Analysis to identify known bad user agents */
-  public static class UserAgentBlocklistAnalysis
-      extends PTransform<PCollection<Event>, PCollection<Alert>> implements DocumentingTransform {
-    private static final long serialVersionUID = 1L;
-
-    private final String monitoredResource;
-    private final Boolean enableIprepdDatastoreExemptions;
-    private final String iprepdDatastoreExemptionsProject;
-    private final String uaBlocklistPath;
-
-    private PCollectionView<Map<String, Boolean>> natView = null;
-    private final HeuristicMetrics metrics;
-
-    private Logger log;
-
-    /**
-     * Initialize new {@link UserAgentBlocklistAnalysis}
-     *
-     * @param toggles {@link HTTPRequestToggles}
-     * @param enableIprepdDatastoreExemptions True to enable datastore exemptions
-     * @param iprepdDatastoreExemptionsProject Project to look for datastore entities in
-     * @param natView Use {@link DetectNat} view, or null to disable
-     */
-    public UserAgentBlocklistAnalysis(
-        HTTPRequestToggles toggles,
-        Boolean enableIprepdDatastoreExemptions,
-        String iprepdDatastoreExemptionsProject,
-        PCollectionView<Map<String, Boolean>> natView) {
-      monitoredResource = toggles.getMonitoredResource();
-      this.enableIprepdDatastoreExemptions = enableIprepdDatastoreExemptions;
-      this.iprepdDatastoreExemptionsProject = iprepdDatastoreExemptionsProject;
-      this.natView = natView;
-      uaBlocklistPath = toggles.getUserAgentBlocklistPath();
-      log = LoggerFactory.getLogger(UserAgentBlocklistAnalysis.class);
-      metrics = new HeuristicMetrics(UserAgentBlocklistAnalysis.class.getName());
-    }
-
-    /** {@inheritDoc} */
-    public String getTransformDoc() {
-      return new String(
-          "Alert if client makes request with user agent that matches entry in blocklist.");
-    }
-
-    @Override
-    public PCollection<Alert> expand(PCollection<Event> input) {
-      if (natView == null) {
-        // If natView was not set then we just create an empty view for use as the side input
-        natView = DetectNat.getEmptyView(input.getPipeline());
-      }
-      return input
-          .apply(
-              "extract agent and source",
-              ParDo.of(
-                  new DoFn<Event, KV<String, String>>() {
-                    private static final long serialVersionUID = 1L;
-
-                    @ProcessElement
-                    public void processElement(ProcessContext c) {
-                      Normalized n = c.element().getNormalized();
-
-                      String sourceAddress = n.getSourceAddress();
-                      if (sourceAddress == null) {
-                        return;
-                      }
-
-                      String ua = n.getUserAgent();
-                      if (ua == null) {
-                        return;
-                      }
-                      // As an optimization, anything resembling a Firefox user agent we will
-                      // just exclude from further analysis in this transform.
-                      if (ua.contains("Firefox/")) {
-                        return;
-                      }
-                      c.output(KV.of(sourceAddress, ua));
-                    }
-                  }))
-          .apply("distinct agent and source", Distinct.<KV<String, String>>create())
-          .apply(
-              "isolate matching agents",
-              ParDo.of(
-                  new DoFn<KV<String, String>, KV<String, String>>() {
-                    private static final long serialVersionUID = 1L;
-
-                    private Pattern uaRegex;
-
-                    @Setup
-                    public void setup() throws IOException {
-                      ArrayList<String> in = FileUtil.fileReadLines(uaBlocklistPath);
-                      uaRegex = Pattern.compile(String.join("|", in));
-                    }
-
-                    @ProcessElement
-                    public void processElement(ProcessContext c) {
-                      if (uaRegex.matcher(c.element().getValue()).matches()) {
-                        c.output(c.element());
-                      }
-                    }
-                  }))
-          .apply(GroupByKey.<String, String>create())
-          .apply(
-              "user agent blocklist analysis",
-              ParDo.of(
-                      new DoFn<KV<String, Iterable<String>>, Alert>() {
-                        private static final long serialVersionUID = 1L;
-
-                        @ProcessElement
-                        public void processElement(ProcessContext c, BoundedWindow w) {
-                          Map<String, Boolean> nv = c.sideInput(natView);
-
-                          String saddr = c.element().getKey();
-                          // Iterable user agent list not currently used here, could probably be
-                          // included in the alert metadata though.
-
-                          Boolean isNat = nv.get(saddr);
-                          if (isNat != null && isNat) {
-                            log.info(
-                                "{}: detectnat: skipping result emission for {}",
-                                w.toString(),
-                                saddr);
-                            metrics.natDetected();
-                            return;
-                          }
-
-                          Alert a = new Alert();
-                          a.setSummary(
-                              String.format(
-                                  "%s httprequest useragent_blocklist %s",
-                                  monitoredResource, saddr));
-                          a.setCategory("httprequest");
-                          a.setSubcategory("useragent_blocklist");
-                          a.addMetadata(AlertMeta.Key.SOURCEADDRESS, saddr);
-
-                          try {
-                            if (enableIprepdDatastoreExemptions) {
-                              IprepdIO.addMetadataIfIpIsExempt(
-                                  saddr, a, iprepdDatastoreExemptionsProject);
-                            }
-                          } catch (IOException exc) {
-                            log.error("error checking iprepd exemptions: {}", exc.getMessage());
-                            return;
-                          }
-
-                          a.setNotifyMergeKey(
-                              String.format("%s useragent_blocklist", monitoredResource));
-                          a.addMetadata(
-                              AlertMeta.Key.WINDOW_TIMESTAMP,
-                              (new DateTime(w.maxTimestamp())).toString());
-
-                          if (!a.hasCorrectFields()) {
-                            throw new IllegalArgumentException(
-                                "alert has invalid field configuration");
-                          }
-                          c.output(a);
-                        }
-                      })
-                  .withSideInputs(natView));
-    }
-  }
-
-  /**
-   * Transform for detection of a single source making excessive requests of a specific endpoint
-   * path solely.
-   *
-   * <p>Generates alerts where the request profile violates path thresholds specified in the
-   * endpointAbusePath pipeline option configuration.
-   */
-  public static class EndpointAbuseAnalysis
-      extends PTransform<PCollection<KV<String, ArrayList<String>>>, PCollection<Alert>>
-      implements DocumentingTransform {
-    private static final long serialVersionUID = 1L;
-
-    private Logger log;
-
-    private final EndpointAbuseEndpointInfo[] endpoints;
-    private final String monitoredResource;
-    private final Boolean enableIprepdDatastoreExemptions;
-    private final Boolean varianceSupportingOnly;
-    private final String[] customVarianceSubstrings;
-    private final String iprepdDatastoreExemptionsProject;
-    private final Integer suppressRecovery;
-    private final Long sessionGapDurationMinutes;
-    private final Long alertSuppressionDurationSeconds;
-
-    /** Internal class for configured endpoints in EPA */
-    public static class EndpointAbuseEndpointInfo implements Serializable {
-      private static final long serialVersionUID = 1L;
-
-      /** Request method */
-      public String method;
-      /** Request path */
-      public String path;
-      /** Threshold */
-      public Integer threshold;
-    }
-
-    /** Internal class for endpoint abuse state */
-    public static class EndpointAbuseState implements Serializable {
-      private static final long serialVersionUID = 1L;
-
-      /** Remote address */
-      public String remoteAddress;
-      /** Request count */
-      public Integer count;
-      /** Timestamp */
-      public Instant timestamp;
-    }
-
-    /**
-     * Static initializer for {@link EndpointAbuseAnalysis}
-     *
-     * @param toggles {@link HTTPRequestToggles}
-     * @param enableIprepdDatastoreExemptions True to enable datastore exemptions
-     * @param iprepdDatastoreExemptionsProject Project to look for datastore entities in
-     */
-    public EndpointAbuseAnalysis(
-        HTTPRequestToggles toggles,
-        Boolean enableIprepdDatastoreExemptions,
-        String iprepdDatastoreExemptionsProject) {
-      log = LoggerFactory.getLogger(EndpointAbuseAnalysis.class);
-
-      monitoredResource = toggles.getMonitoredResource();
-      this.enableIprepdDatastoreExemptions = enableIprepdDatastoreExemptions;
-      this.iprepdDatastoreExemptionsProject = iprepdDatastoreExemptionsProject;
-      varianceSupportingOnly = toggles.getEndpointAbuseExtendedVariance();
-      suppressRecovery = toggles.getEndpointAbuseSuppressRecovery();
-      customVarianceSubstrings = toggles.getEndpointAbuseCustomVarianceSubstrings();
-      sessionGapDurationMinutes = toggles.getSessionGapDurationMinutes();
-      alertSuppressionDurationSeconds = toggles.getAlertSuppressionDurationSeconds();
-
-      String[] cfgEndpoints = toggles.getEndpointAbusePath();
-      endpoints = new EndpointAbuseEndpointInfo[cfgEndpoints.length];
-      for (int i = 0; i < cfgEndpoints.length; i++) {
-        String[] parts = cfgEndpoints[i].split(":");
-        if (parts.length != 3) {
-          throw new IllegalArgumentException(
-              "invalid format for abuse endpoint path, must be <int>:<method>:<path>");
-        }
-        EndpointAbuseEndpointInfo ninfo = new EndpointAbuseEndpointInfo();
-        ninfo.threshold = Integer.parseInt(parts[0]);
-        ninfo.method = parts[1];
-        ninfo.path = parts[2];
-        endpoints[i] = ninfo;
-      }
-    }
-
-    /** {@inheritDoc} */
-    public String getTransformDoc() {
-      String buf = null;
-      for (int i = 0; i < endpoints.length; i++) {
-        String x =
-            String.format(
-                "%d %s requests for %s.",
-                endpoints[i].threshold, endpoints[i].method, endpoints[i].path);
-        if (buf == null) {
-          buf = x;
-        } else {
-          buf += " " + x;
-        }
-      }
-      return String.format(
-          "Clients are sessionized by address, where a session ends after "
-              + "%d minutes of inactivity. An alert is generated if a client is observed "
-              + "making repeated requests to configured endpoints without requesting other forms "
-              + "of content from the site. %s",
-          sessionGapDurationMinutes, buf);
-    }
-
-    @Override
-    public PCollection<Alert> expand(PCollection<KV<String, ArrayList<String>>> input) {
-      return input
-          .apply(GroupByKey.<String, ArrayList<String>>create())
-          .apply(
-              "analyze per-client",
-              ParDo.of(
-                  new DoFn<KV<String, Iterable<ArrayList<String>>>, KV<String, Alert>>() {
-                    private static final long serialVersionUID = 1L;
-
-                    @ProcessElement
-                    public void processElement(ProcessContext c, BoundedWindow w) {
-                      String remoteAddress = c.element().getKey();
-                      Iterable<ArrayList<String>> paths = c.element().getValue();
-                      int[] endCounter = new int[endpoints.length];
-                      String userAgent = null;
-                      boolean basicVariance = false;
-                      boolean extendedVariance = false;
-
-                      // Used to track the latest applicable EPA request, so we can use it as the
-                      // alert timestamp if we need to generate an alert.
-                      Instant latestEpaRequest = null;
-
-                      // Count the number of requests in-window for this source that map to
-                      // monitored endpoints. Set a basic variance flag if we see a request
-                      // that was made to something that is not monitored.
-                      for (ArrayList<String> i : paths) {
-                        Integer abIdx = indexEndpoint(i.get(1), i.get(0));
-                        if (abIdx == null) {
-                          if (customVarianceSubstrings != null) {
-                            for (String s : customVarianceSubstrings) {
-                              if (i.get(4).contains(s)) {
-                                basicVariance = true;
-                                extendedVariance = true;
-                              }
-                            }
-                          }
-                          basicVariance = true;
-                          if (considerSupporting(i.get(1))) {
-                            extendedVariance = true;
-                          }
-                          continue;
-                        }
-                        Instant t = Instant.parse(i.get(3));
-                        if (latestEpaRequest == null) {
-                          latestEpaRequest = t;
-                        } else if (t.getMillis() > latestEpaRequest.getMillis()) {
-                          latestEpaRequest = t;
-                        }
-                        // XXX Just pick up the user agent here; with agent variance this could
-                        // result in a different agent being included in the alert than the one
-                        // that was actually associated with the threshold violation, and should
-                        // be fixed.
-                        userAgent = i.get(2);
-                        endCounter[abIdx]++;
-                      }
-
-                      // If extended object variance is enabled, only consider variance if this
-                      // flag has been set. Otherwise we by default consider basic variance to be
-                      // enough.
-                      if (varianceSupportingOnly) {
-                        if (extendedVariance) {
-                          return;
-                        }
-                      } else {
-                        if (basicVariance) {
-                          return;
-                        }
-                      }
-
-                      // If we get here, there was not enough variance present, identify if any
-                      // monitored endpoints have exceeded the threshold and use the one with
-                      // the highest request count
-                      Integer abmaxIndex = null;
-                      int count = -1;
-                      for (int i = 0; i < endpoints.length; i++) {
-                        if (endpoints[i].threshold <= endCounter[i]) {
-                          if (abmaxIndex == null) {
-                            abmaxIndex = i;
-                            count = endCounter[i];
-                          } else {
-                            if (count < endCounter[i]) {
-                              abmaxIndex = i;
-                              count = endCounter[i];
-                            }
-                          }
-                        }
-                      }
-
-                      if (abmaxIndex == null) {
-                        // No requests exceeded threshold
-                        return;
-                      }
-
-                      log.info("{}: emitting alert for {} {}", w.toString(), remoteAddress, count);
-
-                      String compareMethod = endpoints[abmaxIndex].method;
-                      String comparePath = endpoints[abmaxIndex].path;
-
-                      Alert a = new Alert();
-                      a.setTimestamp(latestEpaRequest.toDateTime());
-                      a.setSummary(
-                          String.format(
-                              "%s httprequest endpoint_abuse %s %s %s %d",
-                              monitoredResource, remoteAddress, compareMethod, comparePath, count));
-                      a.setCategory("httprequest");
-                      a.setSubcategory("endpoint_abuse");
-                      a.addMetadata(AlertMeta.Key.SOURCEADDRESS, remoteAddress);
-
-                      try {
-                        if (enableIprepdDatastoreExemptions) {
-                          IprepdIO.addMetadataIfIpIsExempt(
-                              remoteAddress, a, iprepdDatastoreExemptionsProject);
-                        }
-                      } catch (IOException exc) {
-                        log.error("error checking iprepd exemptions: {}", exc.getMessage());
-                        return;
-                      }
-
-                      if (suppressRecovery != null) {
-                        IprepdIO.addMetadataSuppressRecovery(suppressRecovery, a);
-                      }
-
-                      a.addMetadata(AlertMeta.Key.ENDPOINT, comparePath);
-                      a.addMetadata(AlertMeta.Key.METHOD, compareMethod);
-                      a.addMetadata(AlertMeta.Key.COUNT, Integer.toString(count));
-                      a.addMetadata(AlertMeta.Key.USERAGENT, userAgent);
-                      a.setNotifyMergeKey(String.format("%s endpoint_abuse", monitoredResource));
-                      a.addMetadata(
-                          AlertMeta.Key.WINDOW_TIMESTAMP,
-                          (new DateTime(w.maxTimestamp())).toString());
-                      if (!a.hasCorrectFields()) {
-                        throw new IllegalArgumentException("alert has invalid field configuration");
-                      }
-                      c.output(KV.of(remoteAddress, a));
-                    }
-                  }))
-          // Rewindow into global windows so we can use state in the next step
-          //
-          // Ideally we would apply state in the previous step, but this is not currently
-          // supported by DataflowRunner.
-          //
-          // See also https://issues.apache.org/jira/browse/BEAM-2507
-          .apply("endpoint abuse analysis global", new GlobalTriggers<KV<String, Alert>>(5))
-          .apply(ParDo.of(new AlertSuppressorCount(alertSuppressionDurationSeconds)));
-    }
-
-    private Boolean considerSupporting(String path) {
-      if ((path.endsWith(".css"))
-          || (path.endsWith(".js"))
-          || (path.endsWith(".gif"))
-          || (path.endsWith(".jpg"))
-          || (path.endsWith(".ico"))
-          || (path.endsWith(".svg"))
-          || (path.endsWith(".png"))) {
-        return true;
-      }
-      return false;
-    }
-
-    private Integer indexEndpoint(String path, String method) {
-      for (int i = 0; i < endpoints.length; i++) {
-        if ((endpoints[i].method.equals(method)) && (endpoints[i].path.equals(path))) {
-          return i;
-        }
-      }
-      return null;
-    }
-  }
-
-  /**
-   * Composite transform that conducts threshold analysis using the configured threshold modifier
-   */
-  public static class ThresholdAnalysis extends PTransform<PCollection<Event>, PCollection<Alert>>
-      implements DocumentingTransform {
-    private static final long serialVersionUID = 1L;
-
-    private final Double thresholdModifier;
-    private final Double requiredMinimumAverage;
-    private final Long requiredMinimumClients;
-    private final Double clampThresholdMaximum;
-    private final Long requiredMinimumRequestsPerClient;
-    private final String monitoredResource;
-    private final Boolean enableIprepdDatastoreExemptions;
-    private final String iprepdDatastoreExemptionsProject;
-    private PCollectionView<Map<String, Boolean>> natView = null;
-
-    private final HeuristicMetrics metrics;
-    private Logger log;
-
-    /**
-     * Static initializer for {@link ThresholdAnalysis}.
-     *
-     * @param toggles {@link HTTPRequestToggles}
-     * @param enableIprepdDatastoreExemptions True to enable datastore exemptions
-     * @param iprepdDatastoreExemptionsProject Project to look for datastore entities in
-     * @param natView Use {@link DetectNat} view, or null to disable
-     */
-    public ThresholdAnalysis(
-        HTTPRequestToggles toggles,
-        Boolean enableIprepdDatastoreExemptions,
-        String iprepdDatastoreExemptionsProject,
-        PCollectionView<Map<String, Boolean>> natView) {
-      this.thresholdModifier = toggles.getAnalysisThresholdModifier();
-      this.requiredMinimumAverage = toggles.getRequiredMinimumAverage();
-      this.requiredMinimumClients = toggles.getRequiredMinimumClients();
-      this.requiredMinimumRequestsPerClient = toggles.getRequiredMinimumRequestsPerClient();
-      this.clampThresholdMaximum = toggles.getClampThresholdMaximum();
-      this.monitoredResource = toggles.getMonitoredResource();
-      this.enableIprepdDatastoreExemptions = enableIprepdDatastoreExemptions;
-      this.iprepdDatastoreExemptionsProject = iprepdDatastoreExemptionsProject;
-      this.natView = natView;
-      this.metrics = new HeuristicMetrics(ThresholdAnalysis.class.getName());
-      log = LoggerFactory.getLogger(ThresholdAnalysis.class);
-    }
-
-    /** {@inheritDoc} */
-    public String getTransformDoc() {
-      return String.format(
-          "Alert if a single source address makes more than %.2f times the calculated"
-              + " mean request rate for all clients within a 1 minute window.",
-          thresholdModifier);
-    }
-
-    @Override
-    public PCollection<Alert> expand(PCollection<Event> col) {
-      if (natView == null) {
-        // If natView was not set then we just create an empty view for use as the side input
-        natView = DetectNat.getEmptyView(col.getPipeline());
-      }
-
-      // Count per source address
-      PCollection<KV<String, Long>> clientCounts =
-          col.apply(
-                  "calculate per client count",
-                  ParDo.of(
-                      new DoFn<Event, String>() {
-                        private static final long serialVersionUID = 1L;
-
-                        @ProcessElement
-                        public void processElement(ProcessContext c) {
-                          Normalized n = c.element().getNormalized();
-                          if (n.getSourceAddress() == null) {
-                            return;
-                          }
-                          c.output(n.getSourceAddress());
-                        }
-                      }))
-              .apply(Count.<String>perElement());
-
-      // For each client, extract the request count
-      PCollection<Long> counts = clientCounts.apply("extract counts", Values.<Long>create());
-
-      // Obtain statistics on the client count population for use as a side input
-      final PCollectionView<Stats.StatsOutput> wStats = Stats.getView(counts);
-
-      return clientCounts
-          .apply(
-              "filter insignificant",
-              ParDo.of(
-                  new DoFn<KV<String, Long>, KV<String, Long>>() {
-                    private static final long serialVersionUID = 1L;
-
-                    @ProcessElement
-                    public void processElement(ProcessContext c) {
-                      if (c.element().getValue() >= requiredMinimumRequestsPerClient) {
-                        c.output(c.element());
-                      }
-                    }
-                  }))
-          .apply(
-              "apply thresholds",
-              ParDo.of(
-                      new DoFn<KV<String, Long>, Alert>() {
-                        private static final long serialVersionUID = 1L;
-
-                        @ProcessElement
-                        public void processElement(ProcessContext c, BoundedWindow w) {
-                          Stats.StatsOutput sOutput = c.sideInput(wStats);
-                          Long uc = sOutput.getTotalElements();
-                          Map<String, Boolean> nv = c.sideInput(natView);
-
-                          Double cMean = sOutput.getMean();
-
-                          if (uc < requiredMinimumClients) {
-                            return;
-                          }
-
-                          if (cMean < requiredMinimumAverage) {
-                            return;
-                          }
-
-                          if ((clampThresholdMaximum != null) && (cMean > clampThresholdMaximum)) {
-                            cMean = clampThresholdMaximum;
-                          }
-
-                          if (c.element().getValue() >= (cMean * thresholdModifier)) {
-                            Boolean isNat = nv.get(c.element().getKey());
-                            if (isNat != null && isNat) {
-                              log.info(
-                                  "{}: detectnat: skipping result emission for {}",
-                                  w.toString(),
-                                  c.element().getKey());
-                              metrics.natDetected();
-                              return;
-                            }
-                            log.info(
-                                "{}: emitting alert for {}", w.toString(), c.element().getKey());
-                            Alert a = new Alert();
-                            a.setSummary(
-                                String.format(
-                                    "%s httprequest threshold_analysis %s %d",
-                                    monitoredResource,
-                                    c.element().getKey(),
-                                    c.element().getValue()));
-                            a.setCategory("httprequest");
-                            a.setSubcategory("threshold_analysis");
-                            a.addMetadata(AlertMeta.Key.SOURCEADDRESS, c.element().getKey());
-
-                            try {
-                              if (enableIprepdDatastoreExemptions) {
-                                IprepdIO.addMetadataIfIpIsExempt(
-                                    c.element().getKey(), a, iprepdDatastoreExemptionsProject);
-                              }
-                            } catch (IOException exc) {
-                              log.error("error checking iprepd exemptions: {}", exc.getMessage());
-                              return;
-                            }
-
-                            a.addMetadata(AlertMeta.Key.MEAN, sOutput.getMean().toString());
-                            a.addMetadata(AlertMeta.Key.COUNT, c.element().getValue().toString());
-                            a.addMetadata(
-                                AlertMeta.Key.THRESHOLD_MODIFIER, thresholdModifier.toString());
-                            a.setNotifyMergeKey(
-                                String.format("%s threshold_analysis", monitoredResource));
-                            a.addMetadata(
-                                AlertMeta.Key.WINDOW_TIMESTAMP,
-                                (new DateTime(w.maxTimestamp())).toString());
-                            if (!a.hasCorrectFields()) {
-                              throw new IllegalArgumentException(
-                                  "alert has invalid field configuration");
-                            }
-                            c.output(a);
-                          }
-                        }
-                      })
-                  .withSideInputs(wStats, natView));
-    }
-  }
-
-  /**
-   * Transform for detection of a single source making a sequence of requests at a speed faster than
-   * what we expect from a normal user.
-   *
-   * <p>Generates alerts where the request profile violates path thresholds specified in the
-   * endpointAbusePath pipeline option configuration.
-   */
-  public static class EndpointSequenceAbuse
-      extends PTransform<PCollection<Event>, PCollection<Alert>> implements DocumentingTransform {
-    private static final long serialVersionUID = 1L;
-
-    private Logger log;
-
-    private final EndpointSequenceAbuseTimingInfo[] endpointPatterns;
-    private final String monitoredResource;
-    private final Boolean enableIprepdDatastoreExemptions;
-    private final String iprepdDatastoreExemptionsProject;
-    private final Integer suppressRecovery;
-    private PCollectionView<Map<String, Boolean>> natView = null;
-    private final HeuristicMetrics metrics;
-
-    /** Internal class for configured endpoints in EPA */
-    public static class EndpointSequenceAbuseTimingInfo implements Serializable {
-      private static final long serialVersionUID = 1L;
-
-      public Integer threshold;
-
-      public String firstMethod;
-      public String firstPath;
-
-      public Integer deltaMs;
-
-      public String secondMethod;
-      public String secondPath;
-
-      /**
-       * Convert configuration to String
-       *
-       * @return Class parameters as a string
-       */
-      public String toString() {
-        return String.format(
-            "%d:%s:%s:%d:%s:%s",
-            threshold, firstMethod, firstPath, deltaMs, secondMethod, secondPath);
-      }
-    }
-
-    /**
-     * Static initializer for {@link EndpointAbuseAnalysis}
-     *
-     * @param toggles {@link HTTPRequestToggles}
-     * @param enableIprepdDatastoreExemptions True to enable datastore exemptions
-     * @param iprepdDatastoreExemptionsProject Project to look for datastore entities in
-     * @param natView Use {@link DetectNat} view, or null to disable
-     */
-    public EndpointSequenceAbuse(
-        HTTPRequestToggles toggles,
-        Boolean enableIprepdDatastoreExemptions,
-        String iprepdDatastoreExemptionsProject,
-        PCollectionView<Map<String, Boolean>> natView) {
-      log = LoggerFactory.getLogger(EndpointSequenceAbuse.class);
-      monitoredResource = toggles.getMonitoredResource();
-      this.enableIprepdDatastoreExemptions = enableIprepdDatastoreExemptions;
-      this.iprepdDatastoreExemptionsProject = iprepdDatastoreExemptionsProject;
-      suppressRecovery = toggles.getEndpointSequenceAbuseSuppressRecovery();
-      this.natView = natView;
-      metrics = new HeuristicMetrics(EndpointSequenceAbuse.class.getName());
-
-      String[] cfgEndpoints = toggles.getEndpointSequenceAbusePatterns();
-      endpointPatterns = new EndpointSequenceAbuseTimingInfo[cfgEndpoints.length];
-      for (int i = 0; i < cfgEndpoints.length; i++) {
-        String[] parts = cfgEndpoints[i].split(":");
-        if (parts.length != 6) {
-          throw new IllegalArgumentException(
-              "invalid format for abuse endpoint timing, must be <int>:<method>:<path>:<int>:<method>:<path>");
-        }
-        EndpointSequenceAbuseTimingInfo ninfo = new EndpointSequenceAbuseTimingInfo();
-        ninfo.threshold = Integer.parseInt(parts[0]);
-        ninfo.firstMethod = parts[1];
-        ninfo.firstPath = parts[2];
-        ninfo.deltaMs = Integer.parseInt(parts[3]);
-        ninfo.secondMethod = parts[4];
-        ninfo.secondPath = parts[5];
-        endpointPatterns[i] = ninfo;
-      }
-    }
-
-    /** {@inheritDoc} */
-    public String getTransformDoc() {
-      String buf = null;
-      for (int i = 0; i < endpointPatterns.length; i++) {
-        String x =
-            String.format(
-                "%d %s %s requests within %d ms of last %s %s request.",
-                endpointPatterns[i].threshold,
-                endpointPatterns[i].secondMethod,
-                endpointPatterns[i].secondPath,
-                endpointPatterns[i].deltaMs,
-                endpointPatterns[i].firstMethod,
-                endpointPatterns[i].firstPath);
-        if (buf == null) {
-          buf = x;
-        } else {
-          buf += " " + x;
-        }
-      }
-      return String.format(
-          "An alert is generated when a client (identified by ip) makes requests for a sequence of endpoints within a configurable delta thought to be atypical of a normal user. %s",
-          buf);
-    }
-
-    @Override
-    public PCollection<Alert> expand(PCollection<Event> input) {
-      if (natView == null) {
-        // If natView was not set then we just create an empty view for use as the side input
-        natView = DetectNat.getEmptyView(input.getPipeline());
-      }
-      return input
-          .apply(
-              "filter events and key by ip",
-              ParDo.of(
-                  new DoFn<Event, KV<String, Event>>() {
-                    private static final long serialVersionUID = 1L;
-
-                    @ProcessElement
-                    public void processElement(ProcessContext c, BoundedWindow w) {
-                      Event event = c.element();
-                      Normalized n = event.getNormalized();
-                      String sourceAddress = n.getSourceAddress();
-                      String method = n.getRequestMethod();
-                      String path = n.getUrlRequestPath();
-                      if (sourceAddress == null || method == null || path == null) {
-                        return;
-                      }
-                      // only output events if they belong to one of our sequences
-                      if (belongsToSequence(method, path)) {
-                        c.output(KV.of(sourceAddress, event));
-                      }
-                    }
-                  }))
-          .apply(GroupByKey.<String, Event>create())
-          .apply(
-              "analyze per-client",
-              ParDo.of(
-                      new DoFn<KV<String, Iterable<Event>>, Alert>() {
-                        private static final long serialVersionUID = 1L;
-
-                        @ProcessElement
-                        public void processElement(ProcessContext c, BoundedWindow w) {
-                          String remoteAddress = c.element().getKey();
-                          Iterable<Event> events = c.element().getValue();
-
-                          // sort events by timestamp
-                          List<Event> eventList =
-                              StreamSupport.stream(events.spliterator(), false)
-                                  .sorted(
-                                      (e1, e2) -> e1.getTimestamp().compareTo(e2.getTimestamp()))
-                                  .collect(Collectors.toList());
-
-                          int[] violationsCounter = new int[endpointPatterns.length];
-                          Instant[] lastViolationTimestamp = new Instant[endpointPatterns.length];
-                          String[] lastViolationUserAgent = new String[endpointPatterns.length];
-
-                          // used to track last time we saw the first part of each
-                          // request sequence used for timing analysis
-                          Instant[] lastFirstRequest = new Instant[endpointPatterns.length];
-
-                          // for each path
-                          for (Event event : eventList) {
-                            Normalized n = event.getNormalized();
-
-                            // check if its a first item in an endpoint sequence
-                            ArrayList<Integer> indices =
-                                findFirstHalfPatternMatches(
-                                    n.getRequestMethod(), n.getUrlRequestPath());
-
-                            // for any sequence its a part of update the latest timestamp for it
-                            for (Integer m : indices) {
-                              lastFirstRequest[m] = Instant.parse(event.getTimestamp().toString());
-                            }
-
-                            // check if its a second item in an endpoint sequence
-                            ArrayList<Integer> secondIndices =
-                                findSecondHalfPatternMatches(
-                                    n.getRequestMethod(), n.getUrlRequestPath());
-
-                            // for any sequence its the second part of, check the delta and increase
-                            // count if it is
-                            for (Integer m : secondIndices) {
-                              Instant ts = Instant.parse(event.getTimestamp().toString());
-                              if (lastFirstRequest[m] != null) {
-                                if (ts.isBefore(
-                                    lastFirstRequest[m].plus(endpointPatterns[m].deltaMs))) {
-                                  lastViolationUserAgent[m] =
-                                      n.getUserAgent() == null ? "" : n.getUserAgent();
-                                  lastViolationTimestamp[m] = ts;
-                                  violationsCounter[m]++;
-                                }
-                              }
-                            }
-                          }
-
-                          // identify if any monitored endpoints have
-                          // exceeded the threshold and use the one with
-                          // the highest count
-                          Integer abmaxIndex = null;
-                          int count = -1;
-                          for (int i = 0; i < endpointPatterns.length; i++) {
-                            if (endpointPatterns[i].threshold <= violationsCounter[i]) {
-                              if (abmaxIndex == null) {
-                                abmaxIndex = i;
-                                count = violationsCounter[i];
-                              } else {
-                                if (count < violationsCounter[i]) {
-                                  abmaxIndex = i;
-                                  count = violationsCounter[i];
-                                }
-                              }
-                            }
-                          }
-
-                          if (abmaxIndex == null) {
-                            // No requests exceeded threshold
-                            return;
-                          }
-
-                          Map<String, Boolean> nv = c.sideInput(natView);
-                          Boolean isNat = nv.get(remoteAddress);
-                          if (isNat != null && isNat) {
-                            log.info(
-                                "{}: detectnat: skipping result emission for {}",
-                                w.toString(),
-                                remoteAddress);
-                            metrics.natDetected();
-                            return;
-                          }
-
-                          log.info(
-                              "{}: emitting alert for {} {}", w.toString(), remoteAddress, count);
-
-                          String compareFirstMethod = endpointPatterns[abmaxIndex].firstMethod;
-                          String compareFirstPath = endpointPatterns[abmaxIndex].firstPath;
-
-                          Integer compareDelta = endpointPatterns[abmaxIndex].deltaMs;
-
-                          String compareSecondMethod = endpointPatterns[abmaxIndex].secondMethod;
-                          String compareSecondPath = endpointPatterns[abmaxIndex].secondPath;
-
-                          Alert a = new Alert();
-                          a.setTimestamp(lastViolationTimestamp[abmaxIndex].toDateTime());
-                          a.setSummary(
-                              String.format(
-                                  "%s httprequest endpoint_sequence_abuse %s %s:%s:%d:%s:%s %d",
-                                  monitoredResource,
-                                  remoteAddress,
-                                  compareFirstMethod,
-                                  compareFirstPath,
-                                  compareDelta,
-                                  compareSecondMethod,
-                                  compareSecondPath,
-                                  count));
-                          a.setCategory("httprequest");
-                          a.setSubcategory("endpoint_sequence_abuse");
-                          a.addMetadata(AlertMeta.Key.SOURCEADDRESS, remoteAddress);
-
-                          try {
-                            if (enableIprepdDatastoreExemptions) {
-                              IprepdIO.addMetadataIfIpIsExempt(
-                                  remoteAddress, a, iprepdDatastoreExemptionsProject);
-                            }
-                          } catch (IOException exc) {
-                            log.error("error checking iprepd exemptions: {}", exc.getMessage());
-                            return;
-                          }
-
-                          if (suppressRecovery != null) {
-                            IprepdIO.addMetadataSuppressRecovery(suppressRecovery, a);
-                          }
-
-                          a.addMetadata(
-                              AlertMeta.Key.ENDPOINT_PATTERN,
-                              endpointPatterns[abmaxIndex].toString());
-                          a.addMetadata(AlertMeta.Key.COUNT, Integer.toString(count));
-                          a.addMetadata(
-                              AlertMeta.Key.USERAGENT, lastViolationUserAgent[abmaxIndex]);
-                          a.setNotifyMergeKey(
-                              String.format("%s endpoint_sequence_abuse", monitoredResource));
-                          a.addMetadata(
-                              AlertMeta.Key.WINDOW_TIMESTAMP,
-                              (new DateTime(w.maxTimestamp())).toString());
-                          if (!a.hasCorrectFields()) {
-                            throw new IllegalArgumentException(
-                                "alert has invalid field configuration");
-                          }
-                          c.output(a);
-                        }
-                      })
-                  .withSideInputs(natView));
-    }
-
-    /**
-     * Returns the indices of the endpoint sequences that the given method/path match the first part
-     * of the pattern
-     */
-    private ArrayList<Integer> findFirstHalfPatternMatches(String method, String path) {
-      ArrayList<Integer> indices = new ArrayList<Integer>();
-      for (int i = 0; i < endpointPatterns.length; i++) {
-        if ((endpointPatterns[i].firstMethod.equals(method))
-            && (endpointPatterns[i].firstPath.equals(path))) {
-          indices.add(i);
-        }
-      }
-      return indices;
-    }
-
-    /**
-     * Returns the indices of the endpoint sequences that the given method/path match the second
-     * part of the pattern
-     */
-    private ArrayList<Integer> findSecondHalfPatternMatches(String method, String path) {
-      ArrayList<Integer> indices = new ArrayList<Integer>();
-      for (int i = 0; i < endpointPatterns.length; i++) {
-        if ((endpointPatterns[i].secondMethod.equals(method))
-            && (endpointPatterns[i].secondPath.equals(path))) {
-          indices.add(i);
-        }
-      }
-      return indices;
-    }
-
-    /**
-     * Returns true if the method/path combination is part of any sequence (regardless of position)
-     */
-    private Boolean belongsToSequence(String method, String path) {
-      return !(findFirstHalfPatternMatches(method, path).isEmpty()
-          && findSecondHalfPatternMatches(method, path).isEmpty());
-    }
-  }
-
-  /**
-   * Transform for detection of a single source generating errors at a given path pattern. The path
-   * is specified using a regex and an alert is generated if there are more than the threshold
-   * amount of requests matching the path with a status in the 400's.
-   *
-   * <p>Generates alerts where the request profile violates path thresholds specified in the
-   * perEndpointErrorRate pipeline option configuration.
-   */
-  public static class PerEndpointErrorRateAnalysis
-      extends PTransform<PCollection<KV<String, ArrayList<String>>>, PCollection<Alert>>
-      implements DocumentingTransform {
-    private static final long serialVersionUID = 1L;
-
-    private Logger log;
-    private final String monitoredResource;
-    private final Boolean enableIprepdDatastoreExemptions;
-    private final String iprepdDatastoreExemptionsProject;
-    private final EndpointErrorInfo[] endpointInfo;
-    private final Integer suppressRecovery;
-    private final Long sessionGapDurationMinutes;
-    private final Long alertSuppressionDurationSeconds;
-
-    /**
-     * Initializer for {@link PerEndpointErrorRateAnalysis}
-     *
-     * @param toggles {@link HTTPRequestToggles}
-     * @param enableIprepdDatastoreExemptions True to enable datastore exemptions
-     * @param iprepdDatastoreExemptionsProject Project to look for datastore entities in
-     */
-    public PerEndpointErrorRateAnalysis(
-        HTTPRequestToggles toggles,
-        Boolean enableIprepdDatastoreExemptions,
-        String iprepdDatastoreExemptionsProject) {
-      log = LoggerFactory.getLogger(PerEndpointErrorRateAnalysis.class);
-
-      this.monitoredResource = toggles.getMonitoredResource();
-      this.enableIprepdDatastoreExemptions = enableIprepdDatastoreExemptions;
-      this.iprepdDatastoreExemptionsProject = iprepdDatastoreExemptionsProject;
-      this.suppressRecovery = toggles.getPerEndpointErrorRateSuppressRecovery();
-      this.sessionGapDurationMinutes = toggles.getErrorSessionGapDurationMinutes();
-      this.alertSuppressionDurationSeconds =
-          toggles.getPerEndpointErrorRateAlertSuppressionDurationSeconds();
-
-      String[] cfgEndpoints = toggles.getPerEndpointErrorRatePaths();
-
-      this.endpointInfo = new EndpointErrorInfo[cfgEndpoints.length];
-
-      for (int i = 0; i < cfgEndpoints.length; i++) {
-        String[] parts = cfgEndpoints[i].split(":");
-        if (parts.length != 3) {
-          throw new IllegalArgumentException(
-              "invalid format for per endpoint error rate analysis, must be <int>:<method>:<path>");
-        }
-        EndpointErrorInfo ninfo = new EndpointErrorInfo();
-        ninfo.threshold = Integer.parseInt(parts[0]);
-        ninfo.method = parts[1];
-        ninfo.path = Pattern.compile(parts[2]);
-
-        endpointInfo[i] = ninfo;
-      }
-    }
-
-    /** Internal class for configured endpoints in PEERA */
-    public static class EndpointErrorInfo implements Serializable {
-      private static final long serialVersionUID = 1L;
-
-      /** Request method */
-      public String method;
-
-      /** Path (pattern) */
-      public Pattern path;
-
-      /** Threshold */
-      public Integer threshold;
-
-      /** Interval */
-      public Long interval;
-
-      /** Returns true if an event matches the endpoint for both method and path */
-      boolean matchesEvent(String eventMethod, String eventPath) {
-        return method.equals(eventMethod) && path.matcher(eventPath).matches();
-      }
-    }
-
-    /** Internal class to keep track of current state for a given endpoint rule for this key */
-    public static class EndpointErrorState implements Serializable {
-      private static final long serialVersionUID = 1L;
-      int errorCounter;
-      Instant mostRecentError;
-      String userAgent;
-
-      /**
-       * Increments the counter and if this is the most recent event seen update the timestamp and
-       * use this event's useragent for any alerts generated
-       */
-      void update(Instant timestamp, String userAgent) {
-        errorCounter++;
-        if (mostRecentError == null || mostRecentError.isBefore(timestamp)) {
-          mostRecentError = timestamp;
-          this.userAgent = userAgent;
-        }
-      }
-    }
-
-    /** {@inheritDoc} */
-    public String getTransformDoc() {
-      String buf = null;
-      for (int i = 0; i < endpointInfo.length; i++) {
-        String x =
-            String.format(
-                "%d errors to endpoints matching %s %s.",
-                endpointInfo[i].threshold, endpointInfo[i].method, endpointInfo[i].path);
-        if (buf == null) {
-          buf = x;
-        } else {
-          buf += " " + x;
-        }
-      }
-      return String.format(
-          "Clients are sessionized by address, where a session ends after "
-              + "%d minutes of inactivity. An alert is generated if a client is observed "
-              + "making repeated requests to configured endpoints that result in higher "
-              + "amount of errors. %s",
-          sessionGapDurationMinutes, buf);
-    }
-
-    @Override
-    public PCollection<Alert> expand(PCollection<KV<String, ArrayList<String>>> input) {
-      return input
-          .apply(GroupByKey.<String, ArrayList<String>>create())
-          .apply(
-              "analyze per-client",
-              ParDo.of(
-                  new DoFn<KV<String, Iterable<ArrayList<String>>>, KV<String, Alert>>() {
-                    private static final long serialVersionUID = 1L;
-
-                    @ProcessElement
-                    public void processElement(ProcessContext c, BoundedWindow w) {
-                      String remoteAddress = c.element().getKey();
-                      Iterable<ArrayList<String>> errors = c.element().getValue();
-                      EndpointErrorState[] state = new EndpointErrorState[endpointInfo.length];
-                      Arrays.setAll(state, i -> new EndpointErrorState());
-
-                      for (ArrayList<String> e : errors) {
-                        String method = e.get(0);
-                        String path = e.get(1);
-                        for (int i = 0; i < endpointInfo.length; i++) {
-                          if (endpointInfo[i].matchesEvent(method, path)) {
-                            Instant ts = Instant.parse(e.get(3));
-                            String userAgent = e.get(2);
-                            state[i].update(ts, userAgent);
-                          }
-                        }
-                      }
-
-                      // find the endpoint with the highest count and use that to
-                      // generate alert
-                      Integer max = null;
-                      int count = -1;
-                      for (int i = 0; i < endpointInfo.length; i++) {
-                        if (endpointInfo[i].threshold <= state[i].errorCounter) {
-                          if (max == null) {
-                            max = i;
-                            count = state[i].errorCounter;
-                          } else {
-                            if (count < state[i].errorCounter) {
-                              max = i;
-                              count = state[i].errorCounter;
-                            }
-                          }
-                        }
-                      }
-
-                      if (max == null) {
-                        return;
-                      }
-
-                      Alert a = new Alert();
-                      a.setTimestamp(state[max].mostRecentError.toDateTime());
-                      a.setSummary(
-                          String.format(
-                              "%s httprequest per_endpoint_error_rate %s %s %s %d",
-                              monitoredResource,
-                              remoteAddress,
-                              endpointInfo[max].method,
-                              endpointInfo[max].path.pattern(),
-                              state[max].errorCounter));
-                      a.setCategory("httprequest");
-                      a.setSubcategory("per_endpoint_error_rate");
-                      a.addMetadata(AlertMeta.Key.SOURCEADDRESS, remoteAddress);
-
-                      if (enableIprepdDatastoreExemptions) {
-                        try {
-                          IprepdIO.addMetadataIfIpIsExempt(
-                              remoteAddress, a, iprepdDatastoreExemptionsProject);
-                        } catch (IOException exc) {
-                          log.error("error checking iprepd exemptions: {}", exc.getMessage());
-                          return;
-                        }
-                      }
-                      if (suppressRecovery != null) {
-                        IprepdIO.addMetadataSuppressRecovery(suppressRecovery, a);
-                      }
-                      a.addMetadata(AlertMeta.Key.COUNT, String.valueOf(state[max].errorCounter));
-                      a.addMetadata(
-                          AlertMeta.Key.ENDPOINT_PATTERN, endpointInfo[max].path.toString());
-                      a.addMetadata(
-                          AlertMeta.Key.ERROR_THRESHOLD,
-                          String.valueOf(endpointInfo[max].threshold));
-                      a.setNotifyMergeKey(
-                          String.format("%s per_endpoint_error_rate", monitoredResource));
-                      a.addMetadata(
-                          AlertMeta.Key.WINDOW_TIMESTAMP,
-                          (new DateTime(w.maxTimestamp())).toString());
-                      if (!a.hasCorrectFields()) {
-                        throw new IllegalArgumentException("alert has invalid field configuration");
-                      }
-                      c.output(KV.of(remoteAddress, a));
-                    }
-                  }))
-          .apply("per endpoint error analysis global", new GlobalTriggers<KV<String, Alert>>(5))
-          .apply(ParDo.of(new AlertSuppressorCount(alertSuppressionDurationSeconds)));
     }
   }
